pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pytoolkit"
version = "0.2.0"
description = "A collection of helper utilities for Python projects."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT License"}
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
pytoolkit = "pytoolkit.cli:run_cli"
